refactor(guest-agent): extract cli event delivery state
