+++ crates/guest-agent/src/cli/mod.rs
@@ -2,6 +2,7 @@
 
 mod command;
 mod diagnostics;
+mod event_delivery;
 
 pub use command::build_cli_command;
 
@@ -13,6 +14,7 @@ use crate::http::HttpClient;
 use crate::masker::SecretMasker;
 use crate::paths;
 use crate::timing;
+use event_delivery::{AckedEventPrefix, PreparedEvent};
 use guest_common::telemetry::record_sandbox_op;
 use guest_common::{log_info, log_warn};
 use std::collections::HashMap;
@@ -228,43 +230,6 @@ fn setup_codex_chatgpt() -> Result<(), AgentError> {
         log_info!(LOG_TAG, "Codex ChatGPT-OAuth auth.json written");
     }
     result
-}
-
-struct PreparedEvent {
-    sequence: u32,
-    payload: serde_json::Value,
-}
-
-#[derive(Default)]
-struct AckedEventPrefix {
-    next_expected: u32,
-    last_contiguous: Option<u32>,
-    prefix_broken: bool,
-}
-
-impl AckedEventPrefix {
-    fn record_success(&mut self, sequence: u32) {
-        if self.prefix_broken {
-            return;
-        }
-
-        if sequence == self.next_expected {
-            self.last_contiguous = Some(sequence);
-            self.next_expected = sequence.saturating_add(1);
-        } else if sequence > self.next_expected {
-            self.prefix_broken = true;
-        }
-    }
-
-    fn record_failure(&mut self, sequence: u32) {
-        if sequence >= self.next_expected {
-            self.prefix_broken = true;
-        }
-    }
-
-    fn last_contiguous(&self) -> Option<u32> {
-        self.last_contiguous
-    }
 }
 
 /// Summary of Claude Code's terminal `type=result` event.
@@ -922,53 +887,6 @@ mod tests {
         CliFrameworkBehavior::new(env::Framework::ClaudeCode)
             .track_claude_tool_events(&tool_result, &mut tracker);
         assert!(tracker.is_empty());
-    }
-
-    // -----------------------------------------------------------------
-    // AckedEventPrefix
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn acked_event_prefix_advances_on_contiguous_successes() {
-        let mut prefix = AckedEventPrefix::default();
-
-        prefix.record_success(0);
-        prefix.record_success(1);
-        prefix.record_success(2);
-
-        assert_eq!(prefix.last_contiguous(), Some(2));
-    }
-
-    #[test]
-    fn acked_event_prefix_stops_at_first_failed_event() {
-        let mut prefix = AckedEventPrefix::default();
-
-        prefix.record_success(0);
-        prefix.record_failure(1);
-        prefix.record_success(2);
-
-        assert_eq!(prefix.last_contiguous(), Some(0));
-    }
-
-    #[test]
-    fn acked_event_prefix_has_no_watermark_when_first_event_fails() {
-        let mut prefix = AckedEventPrefix::default();
-
-        prefix.record_failure(0);
-        prefix.record_success(1);
-
-        assert_eq!(prefix.last_contiguous(), None);
-    }
-
-    #[test]
-    fn acked_event_prefix_rejects_success_gap() {
-        let mut prefix = AckedEventPrefix::default();
-
-        prefix.record_success(0);
-        prefix.record_success(2);
-        prefix.record_success(3);
-
-        assert_eq!(prefix.last_contiguous(), Some(0));
     }
 
     // -----------------------------------------------------------------
